LXD is Linux only (for now), so be explicit about it - only build LXD code for the linux service.
